fix the type of book id

+++ src/main/java/com/example/demotaskregistration/controller/BookController.java
@@ -54,7 +54,7 @@ public class BookController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<String> deleteBook(@PathVariable("id") Integer id) {
+    public ResponseEntity<String> deleteBook(@PathVariable("id") Long id) {
         try {
             return bookService.delete(id);
         }
